pyproject.toml
[project]
name = "arcagi3-research"
version = "0.1.0"
description = "ARC-AGI research — LLM inference via Azure AI Foundry"
requires-python = ">=3.11"
dependencies = [
    "arc-agi>=0.9.1",
    "openai>=1.50.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4.0",
]

[tool.ruff]
line-length = 100
